// shared/types.d.ts
export interface StripePrice {
    id: string;
    amount: number;
    currency: string;
    credits: number;
    popular?: boolean;
}